# Performance improvement (addresses #15)