--- a/Tests/SwiftPriorityQueueTests/SwiftPriorityQueueTests.swift
+++ b/Tests/SwiftPriorityQueueTests/SwiftPriorityQueueTests.swift
@@ -156,6 +156,26 @@ class SwiftPriorityQueueTests: XCTestCase {
         XCTAssertEqual(expected, actual)
     }
     
+    func testBuildPerformance() {
+        var pq: PriorityQueue<Int> = PriorityQueue<Int>()
+        let input: [Int] = Array((0 ..< 100000))
+        measure {
+            pq = PriorityQueue(ascending: true, startingValues: input)
+        }
+        
+        let actual = Array(pq)
+        XCTAssertEqual(input, actual)
+    }
+    
+    func testPopPerformance() {
+        measure {
+            var pq = PriorityQueue(ascending: true, startingValues: Array(0 ..< 100000))
+            for _ in 0 ..< 100000 {
+                pq.pop()
+            }
+        }
+    }
+    
     static var allTests = [
         ("testCustomOrder", testCustomOrder),
         ("testBasic", testBasic),
@@ -165,6 +185,8 @@ class SwiftPriorityQueueTests: XCTestCase {
         ("testPeek", testPeek),
         ("testRemove", testRemove),
         ("testRemoveAll", testRemoveAll),
-        ("testRemoveLastInHeap", testRemoveLastInHeap)
+        ("testRemoveLastInHeap", testRemoveLastInHeap),
+        ("testBuildPerformance", testBuildPerformance),
+        ("testPopPerformance", testPopPerformance)
         ]
 }