This update implements some lower-level functions to improve the performance of standard heap operations on `SwiftPriorityQueue` and adds performance tests for `init`, `pop`, `push` and `remove`. Hopefully the improvement is sufficient to close #15.

A naive implementation replacing `sink` and `swim` with versions that can be called from inside `Array.withUnsafeMutableBufferPointer(_:)` already gave promising performance for `init`, `pop` and `push`. (See [forked master branch](https://github.com/peteraisher/SwiftPriorityQueue/tree/master). This might represent a good compromise solution if the final code is deemed to be too complex.)

A more in-depth rewrite of the internal code (with inspiration from [libcxx/include/algorithm](https://github.com/google/libcxx/blob/master/include/algorithm)) gave further performance improvements, although the resulting code is less "easy" to understand. In an effort to aid readability, the new code is extensively (some might say excessively!) marked up.

Building a `SwiftPriorityQueue` from an `Array` of 100,000 values went from 17.8 ms to 4.1 ms.
Emptying a `SwiftPriorityQueue` with 100,000 values using `pop()` went from 2.11 s to 0.36 s.
Calling `remove(_:)` on 100 random items from a `SwiftPriorityQueue` with 100,000 values went from 2.36 s to 0.277 s.

Building a `SwiftPriorityQueue` by calling `push(_:)` 100,000 times went from 118 ms to 149 ms. Not certain what happened there! Further investigation may be required, but in any case, the performance improvement in `pop()` more than compensates in normal usage where `pop()` and `push(_:)` are called equally often.
 
<img width="760" alt="Performance Comparison SwiftPriorityQueue" src="https://user-images.githubusercontent.com/212052/65820051-c9920200-e224-11e9-895e-f50c0e833da2.png">