@davecom 2019-09-28
Hi Peter,
Thanks for the extensive write up and contribution. This is all very interesting. It is in effect a total rewrite from scratch. It's interesting that the push() performance went down. I assume the gains here are mostly from reimplementation at a lower level rather than algorithmic changes (it's still a binary heap I presume?), so I would've expected everything to be improved. 

My concern is that merging this would lead to new code that I don't completely understand and would therefore be hard more me to maintain going forward. It's going to a new level of complexity. SwiftPriorityQueue has been around for about 5 years and is used in many real world projects, so if I make the code hard on myself to maintain, the detriment of that could be greater than any performance gain. I'm wondering if there are some gains here that can be gleaned from more minor changes to the existing codebase? Is there an 80/20 rule here?

Why is your version faster? Is it because it is implemented at a lower level with pointers, or because it is using a different algorithm at some point?

Also, more immediately, can you put the performance tests into a separate pull request so I can try to see if I can combine some of the ideas from your version into the mainline version? Thanks in advance.

Thank you so much for all of the hard work here,
David