@peteraisher 2019-09-28
Hi,

My pleasure! It was quite fun to see how far it could be pushed! Yes, over the course of the day it gradually morphed into a total rewrite. Not sure what the issue was with `push(_:)` performance either. There are no algorithmic changes, it's still a binary heap, so I also would've expected improvements across the board. It's possible that the extra overhead from repeatedly entering the `withUnsafeMutablePointer(_:)` closure outweighs the benefits there. I didn't get round to doing proper profiling etc.

I fully understand your concern about maintainability. Luckily there are still pretty big gains to be had just from tweaking things to call `withUnsafeMutablePointer(_:)` instead of working with the `Array`. That gave similar performance gains (still a few times faster) for moderate extra complexity, so maybe this is the way you want to go. The code is [here](https://github.com/peteraisher/SwiftPriorityQueue/blob/master/Sources/SwiftPriorityQueue/SwiftPriorityQueue.swift).

In the meantime, I will create a new PR with just the performance tests.